takeover: Fix header includes
The ioctl.h header has been removed with commit 6495aef5b625b9ddae
("net-snk: Remove module system"), so it can not be included in
the takeover client anymore. Thus remove the include statement (which
is no problem since the takeover client does not use ioctls anyway).
Include unistd.h instead, to avoid that the compiler is printing
out a warning about a missing prototype of the function sbrk().

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>