pci-properties: Enforce all MMIO BARs to be 64K page aligned
QEMU may passthrough a pci device of which BARs are smaller than
64KB(PAGE_SIZE) to guest if the BARs are in an exclusive page.
But these passthru-BARs' mmio page may be shared with other
BARs in guest. This would cause the pci-passthrough fail and mmio
emulation happens in host.

To solve this performance issue, this patch enforces the alignment
of all MMIO BARs allocations to be at least 64K page aligned so that
the mmio page of one BAR would not be shared with other BARs in guest.

Signed-off-by: Yongji Xie <xyjxie@linux.vnet.ibm.com>
Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>